16.04 update
There was a critical bug in supervisor where it didn't restart upon system reboot. A fix has been released; however, it's in the xenial-proposed archive. I'm setting this so that we grab that version of supervisor. Eventually, this will be updated as the package gets into the main xenial release branch.